apps/gemini-chat-image-editor/requirements.txt
streamlit
google-genai
pillow

